day1/02_streamlit_app/requirements.txt
streamlit>=1.24.0
transformers>=4.11.0
torch>=1.9.0
sqlite3
python-dotenv>=0.19.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
accelerate
janome
pyngrok